+++ prism-core/src/main/resources/sql/mysql/prism_get_or_create_entity_type.sql
@@ -1,5 +1,5 @@
 CREATE PROCEDURE %prefix%get_or_create_entity_type
-    (IN `entityType` VARCHAR(25), IN `translationKey` VARCHAR(155), OUT `entityTypeId` SMALLINT)
+    (IN `entityType` VARCHAR(25), IN `translationKey` VARCHAR(155), OUT `entityTypeId` INT)
 BEGIN
     SELECT entity_type_id INTO `entityTypeId` FROM
     %prefix%entity_types WHERE entity_type = `entityType`;
